Moved files around and pytest functioning
