+++ nile_project/nile.py
@@ -20,4 +20,4 @@ SHIPPING_PRICES = {
 def format_price(price):
     return "${0:.2f}".format(price)
 
-print(get_distance(0,0,1,1))
+# print(get_distance(0,0,1,1))
