+++ nile_project/test_nile/test_calculate_money_made.py
@@ -1,0 +1,13 @@
+import sys
+sys.path.append("/Users/RexLAP/PycharmProjects/py_bootcamp/nile_project/")
+from main import * # not importing anything why did it lose access to the path
+import pytest
+
+
+def test_calculate_money_made():
+    # trip1 = Trip(200, driver1, 15)
+    # trip2 = Trip(300, driver2, 40)  
+    # expected_ans = 445 
+    pass
+def test_m():
+    pass
